// DXC_Frontend/src/api/models/articleDto.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * DXC_Core.API
 * OpenAPI spec version: 1.0
 */

export interface ArticleDto {
  publicId?: string
  /** @nullable */
  title?: string | null
  /** @nullable */
  slug?: string | null
  /** @nullable */
  summary?: string | null
  /** @nullable */
  content?: string | null
  /** @nullable */
  coverImagePublicId?: string | null
  /** @nullable */
  coverImageUrl?: string | null
  /** @nullable */
  authorName?: string | null
  viewCount?: number
  /** @nullable */
  publishedAt?: string | null
  thuTu?: number
  isActive?: boolean
  createdAt?: string
  updatedAt?: string
}